tmp: refactor
