+++ js/packages/berty-storybook/shared-components/SimpleModal.tsx
